API Bitti

+++ Azure/Controllers/FilmController.cs
@@ -26,6 +26,20 @@ namespace Azure.Controllers
         [Function("GetFilmler")]
         public async Task<IActionResult> GetFilmler([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "film")] HttpRequest req)
         {
+            string? authHeader = req.Headers.Authorization;
+            if(string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+                return new UnauthorizedObjectResult("Token yok veya yanlis");
+
+            string token = authHeader["Bearer ".Length..].Trim();
+
+            if(!_tokenService.ValidateToken(token))
+                return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] == null)
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 var filmler = await _filmRepo.GetAllFilmlerAsync();
@@ -40,6 +54,20 @@ namespace Azure.Controllers
         [Function("GetFilmById")]
         public async Task<IActionResult> GetFilmById([HttpTrigger(AuthorizationLevel.Anonymous, "get", Route = "film/{id}")] HttpRequest req, int id)
         {
+            string? authHeader = req.Headers.Authorization;
+            if(string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+                return new UnauthorizedObjectResult("Token yok veya yanlis");
+
+            string token = authHeader["Bearer ".Length..].Trim();
+
+            if(!_tokenService.ValidateToken(token))
+                return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] == null)
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 var film = await _filmRepo.GetFilmByIdAsync(id);
@@ -54,6 +82,20 @@ namespace Azure.Controllers
         [Function("DeleteFilmById")]
         public async Task<IActionResult> DeleteFilmById([HttpTrigger(AuthorizationLevel.Anonymous, "delete", Route = "film/{id}")] HttpRequest req, int id)
         {
+            string? authHeader = req.Headers.Authorization;
+            if(string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+                return new UnauthorizedObjectResult("Token yok veya yanlis");
+
+            string token = authHeader["Bearer ".Length..].Trim();
+
+            if(!_tokenService.ValidateToken(token))
+                return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] == "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 var film = await _filmRepo.DeleteFilmByIdAsync(id);
@@ -68,6 +110,20 @@ namespace Azure.Controllers
         [Function("CreateFilm")]
         public async Task<IActionResult> CreateFilm([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = "film")] HttpRequest req)
         {
+            string? authHeader = req.Headers.Authorization;
+            if(string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+                return new UnauthorizedObjectResult("Token yok veya yanlis");
+
+            string token = authHeader["Bearer ".Length..].Trim();
+
+            if(!_tokenService.ValidateToken(token))
+                return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] == "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 using var reader = new StreamReader(req.Body);
@@ -89,6 +145,20 @@ namespace Azure.Controllers
         [Function("UpdateFilm")]
         public async Task<IActionResult> UpdateFilm([HttpTrigger(AuthorizationLevel.Anonymous, "put", Route = "film/{id}")] HttpRequest req, int id)
         {
+            string? authHeader = req.Headers.Authorization;
+            if(string.IsNullOrEmpty(authHeader) || !authHeader.StartsWith("Bearer "))
+                return new UnauthorizedObjectResult("Token yok veya yanlis");
+
+            string token = authHeader["Bearer ".Length..].Trim();
+
+            if(!_tokenService.ValidateToken(token))
+                return new UnauthorizedObjectResult("Gecersiz Token");
+
+            var claims = _tokenService.GetTokenClaims(token);
+
+            if(claims["role"] == "Admin")
+                return new UnauthorizedObjectResult("Yetkisiz Erisim");
+
             try
             {
                 using var reader = new StreamReader(req.Body);
